[rest] `serve.test.ts`

+++ src/rest/handleRequest.ts
@@ -4,9 +4,19 @@ import type { Database } from "../database"
 import * as Db from "../db"
 import { requestJsonObject } from "../utils/requestJsonObject"
 
-async function handleRequest(db: Database, request: Http.IncomingMessage) {
+export async function handleRequest(
+  db: Database,
+  request: Http.IncomingMessage,
+) {
   if (request.url === undefined) {
     throw new Error("[handleRequest] expect request.url")
+  }
+
+  const url = new URL(request.url, `http://${request.headers.host}`)
+  const id = url.pathname.slice(1)
+
+  if (request.method === "GET") {
+    return await Db.get(db, url.pathname)
   }
 
   if (request.headers["content-type"] !== "application/json") {
@@ -15,41 +25,28 @@ async function handleRequest(db: Database, request: Http.IncomingMessage) {
     )
   }
 
-  const url = new URL(request.url, `http://${request.headers.host}`)
-  const id = url.pathname.slice(1)
-
-  switch (request.method) {
-    case "GET": {
-      return await Db.get(db, url.pathname)
-    }
-
-    case "POST": {
-      const json = await requestJsonObject(request)
-      return await Db.create(db, { ...json, "@id": id })
-    }
-
-    case "PUT": {
-      const json = await requestJsonObject(request)
-      const input = dataOmitIdFromJson(json)
-      return await Db.put(db, { ...input, "@id": id })
-    }
-
-    case "PATCH": {
-      const json = await requestJsonObject(request)
-      const input = dataOmitIdFromJson(json)
-      return await Db.patch(db, { ...input, "@id": id })
-    }
-
-    case "DELETE": {
-      const json = await requestJsonObject(request)
-      const input = dataOmitIdFromJson(json)
-      return await Db.delete(db, { ...input, "@id": id })
-    }
-
-    default: {
-      throw new Error(
-        `[handleRequest] unhandled http method: ${request.method}`,
-      )
-    }
+  if (request.method === "POST") {
+    const json = await requestJsonObject(request)
+    return await Db.create(db, { ...json, "@id": id })
   }
+
+  if (request.method === "PUT") {
+    const json = await requestJsonObject(request)
+    const input = dataOmitIdFromJson(json)
+    return await Db.put(db, { ...input, "@id": id })
+  }
+
+  if (request.method === "PATCH") {
+    const json = await requestJsonObject(request)
+    const input = dataOmitIdFromJson(json)
+    return await Db.patch(db, { ...input, "@id": id })
+  }
+
+  if (request.method === "DELETE") {
+    const json = await requestJsonObject(request)
+    const input = dataOmitIdFromJson(json)
+    return await Db.delete(db, { ...input, "@id": id })
+  }
+
+  throw new Error(`[handleRequest] unhandled http method: ${request.method}`)
 }
